Refactor all the code! Makes it a bit easier to work with. First real play with async/await. It appears to work, though I'm probably doing everything wrong. Constructive criticism welcome...!
